Rough first cut of changing tree from recursive to flat structure
To see if it made some of the operations easier. Actually added and tested
tighter restrictions on mutating the tree:
 - Keys can only be reordered
 - (Clapi) types cannot change

The downside is we have to manually check if there's a parent element to add to,
and we have to do some fairly nasty filtering to remove a subtree on delete.

Also, the code just got distinctly more proceedural, and much more like the
Python...